Update `@metamask/announcement-controller` to v4
The `@metamask/announcement-controller` has been updated to v4, which
is the version released in the core monorepo v53 release. The remaining
packages released as part of v53 will be updated in later PRs.

The only breaking change is that the minimum Node.js version was
updated to v16.

Relates to #19271